use memo as payee with these formats that don't have payee
The reasoning here is that the payee rules in the  YNAB app can use the memo information to discover the payee (and category, if specified with the payee configuration). That would not work if the info is provided as memo; hence this seems a good idea. It does require the user to maintain his payees with relevant rules.