added SE Swedbank 2019
This is untested but I'll ask the format contributor to test and respond. Also, this is the third format for filename "export.csv" so the tool won't be able to apply the correct conversion because it only looks at the filename. We need to find a better approach to match files to formats.