Import tree
NB: in a state of works for me on my machine.

But the general idea is specify the bank up front. I found the the pattern matching file finder was
not suited when i could download with any name for my files

Then import a tree as in

Get all budgets and accounts from ynab api and create a structure like

```
imports/
└── My Budget
    ├── AL-budget
    ├── AL-Helle
    ├── Al-jacob
    └── AL-MadKontoen
```

Then walk trough all those direcories and post any bank-transactions into ynab

Finally move the files to a history folder with a similar structure as the import tree, so they are
only processed once.